+++ docenten-api/app/Http/Controllers/Api/CourseTypeController.php
@@ -7,17 +7,21 @@ use App\Models\CourseType;
 use App\Http\Resources\CourseTypeResource;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Gate;
 
 class CourseTypeController extends Controller
 {
 
     public function index()
     {
+        Gate::authorize('viewAny', CourseType::class);
         return CourseTypeResource::collection(CourseType::all());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', CourseType::class);
+
         $validated = $request->validate([
             'name' => [
                 'required',
@@ -35,6 +39,7 @@ class CourseTypeController extends Controller
 
     public function show(CourseType $courseType)
     {
+        Gate::authorize('view', $courseType);
         $courseType->load('courses');
 
         return new CourseTypeResource($courseType);
@@ -42,6 +47,7 @@ class CourseTypeController extends Controller
 
     public function update(Request $request, CourseType $courseType)
     {
+        Gate::authorize('update', $courseType);
         $validated = $request->validate([
             'name' => [
                 'sometimes',
@@ -59,6 +65,8 @@ class CourseTypeController extends Controller
 
     public function destroy(CourseType $courseType)
     {
+        Gate::authorize('delete', $courseType);
+
         $courseType->delete();
 
         return response()->noContent();
